+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Updated shakapaker to 6.5.5
 - Fix a bug: unable to create a recurrent event
 - Fix a bug: unable to create a non-rolling plan
+- Fix a bug: invalid duration for machine/spaces reservations in statistics, when using slots of not 1 hour
 - Fix a bug: invalid month in date format
 - Fix a bug: do not show theme and age-range fields in event form if no options were set
 - Fix a bug: do not show catgory select in plan form if no options were set